Refactor spicecmd
Language specific parameters are now in <simulator>_lang modules
spicecmd property is now defined in simulator specific module
